Extract some functions into base/str.cpp
See #10093
